backend/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
defusedxml==0.7.1
Django==3.2
django-allauth==0.44.0
django-filter==2.4.0
django-rest-auth==0.9.5
django-shortuuidfield==0.1.3
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-writable-nested==0.6.2
dry-rest-permissions==0.1.10
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
mypy-extensions==0.4.3
oauthlib==3.1.0
pathspec==0.8.1
Pillow==8.2.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==20.0.1
python3-openid==3.2.0
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
service-identity==18.1.0
shortuuid==1.0.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.7.4.3
urllib3==1.26.4
zope.interface==5.3.0
